+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All Notable changes to `larablog` will be documented in this file.
 - Replaced references to 'admin' middleware with middleware referenced in config
 - Replaced calls to function 'isUserAdmin()' with reference to function specified in middleware
 
+### Fixed
+- Simplified controller code
+
 ## v1.1.3 2017-01-16
 
 ### Added